+++ src/app/app.module.ts
@@ -34,7 +34,10 @@ import { MenuComponent } from './components/menu/menu.component'
 import { LottieModule } from 'ngx-lottie'
 import { LogoComponent } from './components/logo/logo.component'
 import { LogoItemComponent } from './components/logo/logo-item/logo-item.component'
-import { FooterComponent } from './footer/footer.component'
+import { FooterComponent } from './components/footer/footer.component'
+import { StatisticLogComponent } from './components/statistic-log/statistic-log.component';
+import { UserInfoComponent } from './components/header/user-info/user-info.component';
+import { ClickedOutsideDirective } from './directives/clicked-outside.directive'
 
 // Factory funtion needed ngx-lottie
 export function playerFactory() {
@@ -56,7 +59,10 @@ export function playerFactory() {
 		MenuComponent,
 		LogoComponent,
 		LogoItemComponent,
-		FooterComponent
+		FooterComponent,
+		StatisticLogComponent,
+  UserInfoComponent,
+  ClickedOutsideDirective
 	],
 	imports: [
 		BrowserModule,