+++ src/app/store/statistic/not-sync/statistic-not-sync.select.ts
@@ -1,0 +1,11 @@
+import { RootState } from '../../rootTypes'
+import { notSyncStatisticToDefault } from '../helpers'
+import { NotSyncTypes } from '.'
+
+export const selectNotSyncStatistic = (
+	state: RootState
+): NotSyncTypes.StateItemWithDefaultStatistic[] => {
+	return [
+		...notSyncStatisticToDefault(state.notSyncStatistic, state.categories)
+	]
+}