+++ src/app/store/statistic/sync/statistic-sync.actions.ts
@@ -1,0 +1,13 @@
+import { props, createActionGroup } from '@ngrx/store'
+import * as SyncTypes from './statistic-sync.types'
+
+export const StatisticSyncActions = createActionGroup({
+	source: 'Statistic sync',
+	events: {
+		set: props<{ statistic: SyncTypes.SyncState }>(),
+		add: props<{ statistic: SyncTypes.StateItem }>(),
+		delete: props<SyncTypes.StateItem>()
+
+		// TODO: Create change
+	}
+})