+++ src/app/store/statistic/sync/statistic-sync.types.ts
@@ -1,0 +1,8 @@
+import { Statistic } from 'src/types/Statistic'
+
+export type StateItem = Statistic & {
+	status?: never
+	action?: never
+}
+
+export type SyncState = StateItem[]